Use JMdict_e XML file for dictionary, change dictionary script
Now using an XML file as a dictionary file, as is easier to parse.

This is very much a WIP; at this point I mainly wanted to see how
well I could handle large amounts of data. Much of the code needs
to be moved into a module and properly tested.